backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
aiofiles
markdown
python-dotenv
google-generativeai
